added `max_values_per_facet` opt to GET /evidence
This option was added to limit the number of facet values returned for
every category, name pair. This will reduce the noise seen for facets
with a high cardinality of the value set (e.g. citation, id).

The option is implemented by filtering the response at the REST API
level. The facet computation is not affected.

closes #73