implement basic /api/expressions/{bel}/components
This returns the subject, relationship, and object components of a BEL
expression. The implementation uses bel.rb's ruby parser so it expects
to parse a BEL statement.

It also depends on some BEL::RDF code to convert short-form
relationships into long form. This introduces very undesirable coupling
that we will want to replace with language version-specific
interpretations.

refs #77