support format param for datasets/{id}/evidence
The format parameter allows a lookup of BEL Translator plugins by its
identifier. If the identifier is not found then a 501 Not Implemented
results.

The format parameter setting wins over the "Accept" header if both are
provided. That means you may pass a valid "Accept" header, but if you
include an invalid format parameter you will receive a 501.