add version reporting to OpenBEL API
Manage the canonical version string from the project in the VERSION file
found in the root of the project. All other version conduits use this
file (hence canonical).

Version string and semantic version numbers are availabe in the API with
OpenBEL::Version.to_s and OpenBEL::Version.to_a. Constants are also
available within that module for MAJOR, MINOR, and PATCH. All strings
and arrays here are frozen. Let it go, let it go...can't hold me back
anyway...let it go, let it go! The perfect girl is gone...etc.

Add GET /api/version route that reports API version information. If
'application/json' is accepted or '?format=json' is passed then a
version resource JSON object is the response. If 'text/plain' is accepted
or '?format=text' is passed then a version string is the response.

Reporting version information in bin/ executables if the -v / --version
option is supplied.

closes #91