batch evidence to an array, avoid JRuby enumerator
The JRuby enumerator uses a thread per next object in an enumerator
which proves costly. Hundreds of threads are created (tested with
yourkit) when batch-creating evidence due to the "each_slice(500)" of
the enumerator.

This issue is logged in JRuby:
https://github.com/jruby/jruby/issues/2577

The solution employed was to yield each evidence directly to the block
and batch 500 into an array at a time. This should avoid the OOM
exception received:

ava.lang.OutOfMemoryError: unable to create new native thread

Indeed the thread count was observed to be lower in yourkit.