aggregate on full-text search; avoids Mongo limits
A full-text search filter to /api/evidence with a sort on bel_statement
only used the text index. This means that the bel_statement sort had to
be done in memory.

This reaches the 32 MB sort limit with only several tens of thousands of
documents.

The solution employed here was to use cursored aggregation allowing disk
use for sort stages.

The solution was introduced as an alternative code path if a FTS filter
was included in the HTTP request. Although this did minimize the risk of
regression there is a fair bit of to clean up in the mongo
access layer.

closes #96