use /data as mounted volume for RDF data
mounting one directory for the RDF data allows both the db file and
rdf store directory to be available. previously the db file mounted
as a directory which resulted in an empty directory.

this explains the experience with empty namespace prefixes since the
db file did not contain any namespaces.